- re-enable ktlint - set kotlin target to 1.6 - set kotlin JVM target to 11 - move shared artifactory publish settings to convention plugin - use kotlin-dsl for referencing tasks and task outputs - rm unnecessary 'java-library' plugin from docs subproject
