+++ buildSrc/src/main/kotlin/buildsrc/convention/ktlint.gradle.kts
@@ -1,11 +1,13 @@
-package buildsrc.config
+package buildsrc.convention
+
+import org.jlleitschuh.gradle.ktlint.KtlintExtension
 
 plugins {
-//    id("org.jlleitschuh.gradle.ktlint") version "10.1.0" // TODO re-enable ktlint
+    id("org.jlleitschuh.gradle.ktlint")
 }
 
-//    configure<org.jlleitschuh.gradle.ktlint.KtlintExtension> {
-//        disabledRules.set(
-//            setOf("no-wildcard-imports", "filename")
-//        )
-//    }
+ktlint {
+    disabledRules.set(
+        setOf("no-wildcard-imports", "filename")
+    )
+}
