+++ buildSrc/src/main/kotlin/buildsrc/convention/subproject.gradle.kts
@@ -1,8 +1,12 @@
 package buildsrc.convention
 
+plugins {
+    id("buildsrc.convention.ktlint")
+}
+
 if (project != rootProject) {
-  group = rootProject.group
-  version = rootProject.version
+    group = rootProject.group
+    version = rootProject.version
 }
 
 description = "Common settings for all Warden subprojects"
